Windows port: build all 8 daemons as native .exe via MSYS2 or zig
POSIX shim layer (win32/):
- winposix.h/.cpp: message-queue-free socket/file/time/process shims used by
  every daemon (poll, readv/writev, scandir, clock_gettime, ctime_r,
  fsync, localtime/gmtime re-entrant helpers, inet helpers). Guard MSG /
  WAIT_TIMEOUT so mingw headers keep working; no afunix.h; lseek/getuid
  macros; C++ mkdir/ftruncate/inet_aton overloads.
- posix/sys shims: un.h (Linux-layout sockaddr_un), poll.h (nfds_t),
  time.h (BSD timer macros), new uio.h (struct iovec).
- wperf.c: portable replacement for the GAS-only cnet/perf/*.s files
  (crc32 + base64, differential-tested by tools/test-wperf.sh).
- zig-only stubs: wpcre.cpp (pcre glue without libpcre), wpopenssl.cpp
  (OpenSSL API stubs), zigonly/openssl headers.

Game source fixes (all no-ops on Linux):
- LLP64 correctness: marshal byteorder/pop_byte overloads, pointer hash
  fold, (int)(intptr_t) casts, off_t -> long long in db/tranlog paths.
- Threading: pthread_sigmask guarded out on Windows (unsupported).
- gs on Windows: no fork(); restart runs detached, multi-server argv
  spawn becomes one-server-per-exe with a warning.
- Storage backup via xcopy on Windows; task headers take server paths
  and real filename casing; daemon __USER__/version printfs hardened.

Build system:
- tools/ci-build-win.sh: native (MSYS2 mingw64, real pcre/crypto) and
  zig (x86_64-windows-gnu cross, stub pcre/crypto) modes; per-step logs
  in .ci-build-win/; stages 8 PE exes + MANIFEST into dist-win/.
- Daemon Makefiles take PCRELIB/CRYPTOLIB/DLLIB ?= overrides.
- .github/workflows/windows.yml: native + zig CI jobs, PE verification,
  artifacts; releases only on manual runs.
- .gitattributes pins LF for scripts/makefiles (CRLF-safe checkout).

+++ cnet/gamed/gsp_if.cpp
@@ -1061,7 +1061,7 @@ void * autoWakeUp(void *)
 
 void * cmdDispatcher(void * tmp)
 {
-	int index = (int)tmp;
+	int index = (int)(intptr_t)tmp;
 	std::deque<user_cmd_t> tmpList;
 	int tustat_time = 0;
 	while(1)
